+++ resources/lang/fi/passwords.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'password' => 'Salasanan tulee olla vähintään kuusi merkkiä pitkä ja sen tulee täsmätä salasanan vahvistukseen.',
+    'reset'    => 'Salasanasi on palautettu!',
+    'sent'     => 'Salasanan palautuslinkki on lähetetty sinulle sähköpostitse',
+    'token'    => 'Salasanan palautusvarmenne ei kelpaa.',
+    'user'     => 'Kyseisellä sähköpostiosoitteella ei löydy käyttäjää',
+];