+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added endpoint to create categories
 - Added file upload option in file section
 - Added endpoint to check if username exists
+- Added finnish translation
 
 ## v0.8 (2018-10-18)